+++ src/components/DataViewer.jsx
@@ -444,7 +444,8 @@ class DataViewer extends Component {
 						<ListItemText primary={`${dataset.title  } - ${  dataset.creator.capitalize()}`}/>
 					</ListItem>);
 				}
-				else if (dataset.format === "shapefile") {
+				else if (dataset.format.toLowerCase() === "shapefile" || dataset.format.toLowerCase() === "raster"
+					|| dataset.format.toLowerCase().includes("geotif")) {
 					return (<ListItem button
 									  onClick={() => this.onClickDataset(dataset.id)}
 									  selected={dataset.id === this.state.selectedDataset.id}>
@@ -547,7 +548,9 @@ class DataViewer extends Component {
 				);
 			}
 			// right column
-			if (this.state.selectedDatasetFormat === "shapefile") {
+			if (this.state.selectedDatasetFormat.toLowerCase() === "shapefile"
+				|| this.state.selectedDatasetFormat.toLowerCase() === "raster"
+				|| this.state.selectedDatasetFormat.toLowerCase().includes("geotif")) {
 				right_column =
 					(<div>
 						<Typography variant="h6">Map</Typography>